refactor(core): extract rust dispatch subsystem from edges/extract
Lift the #200/#207/#208 dispatch-edge synthesis — the closed
conservative handler-call recognizer and the dispatch-fact emitters —
out of extract/mod.rs into extract/rust_dispatch.rs (681 lines). This is
the densest remaining cluster: result_handler_calls and its closed
value-adapter recognizer, arm_rebinds_local, the enum-variant key
helpers, dispatch_fact, rust_dispatch_handle_facts.

rust_edges reaches it through four pub(super) entry points
(enum_variant_key, dispatch_fact, scoped_identifier_in_value_position,
rust_dispatch_handle_facts), qualified as rust_dispatch::*; everything
else is module-private. Pure code movement — logic unchanged, so the
false-edge-is-a-bug invariants are preserved.

extract/mod.rs drops 1490 -> 818 lines. fmt, clippy --all-targets, and
nextest -p rag-rat-core (868 tests, incl. all 19 dispatch_* tests) green.
